refactor: Move functions to dedicated file
Only a single function for now as a sort of proof-of-concept.
